Refactor accumulateOde with bug fix
The god function accumulateOde had a bug in it when using delta while
integrating forward. This bug did not affect any of the current top
level functions. To fix this and prevent problems in the future, the
function was refactored into four different functions, representing the
four combinations of forward/reverse and continuous/discrete.