Introduced m.y(t,x,u), an arbitrary output function
The internals of Kronecker no longer assume that the output is a linear combination of states and inputs. Mass action models still have outputs defined this way, but symbolic models have been changed to allow for arbitrary functions of state and input. THe output function cannot currently be a function of parameters.