Implemented events and introduced observations
As introduced in issue #17, ODE events have been implemented. This was done by creating a new type of object "observations" and changing all the Simulate functions to use the new observations. Also in this commit is great rearraignment of how ODE solutions are passed around, similar to what was proposed in issue #16.