ode15sf: faster LU decomposition/inverse for sensitivity systems
The special form of the Jacobian for sensitivity systems allows one to only take the LU decomposition of the state portion of the Jacobian and decreases the amount of work required to calculate the inverse of the system. This code exploits that structure to speed up the linear solves in the code.