Change to steady state method: iterative integration
The original steady state method relies on comparing the value of the ODE RHS function against the tolerances, which is problematic because the ODE RHS can be very unstable, producing relatively large time derivatives even when the states are sufficiently close to their steady state. The result is that many simulations require a finely-tuned choice of TimeScale -- not too small, or the steady state isn't reached, but also not too large, or the integrator will stall. This can be particularly annoying when running multiple simulations, each of which might need its own tuned TimeScale value.

This new method repeatedly advances the integration by the TimeScale interval, comparing the starting and ending values of the states against their tolerances to determine when a steady state has been reached. The key advantage of this method is that choosing a large value for the TimeScale causes few problems, due to the adaptive time step size of the integrator allowing the steps to grow very large once the steady state is approached.  It also makes the TimeScale more intuitive, since now it is literally the time over which the states should change less than the tolerances when they are at steady state. It also makes it easy to implement a check against no steady state (not implemented yet, to be clear) -- simply limit the number of times the interval is simulated.